% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate_functions.R
\name{aggregate_geneSet}
\alias{aggregate_geneSet}
\title{Aggregate Gene Set Scores}
\usage{
aggregate_geneSet(
  geneList,
  geneSet,
  method = c("mean", "median", "meanabs", "meansqr", "maxmean", "ks_orig", "ks_weighted",
    "ks_pos_neg_sum", "local_fdr", "sign_test", "rank_sum", "custom")
)
}
\arguments{
\item{geneList}{A matrix of genes by models, with each column representing a true or null model.}

\item{geneSet}{A vector containing names of genes in the gene set of interest.}

\item{method}{A character string specifying the method to use for aggregation.
Options include 'mean', 'median', 'meanabs', 'meansqr', 'maxmean',
'sig_n', 'sign_test', 'rank_sum', 'ks_orig', 'ks_weighted', 'ks_sum',
and 'locfdr'. Default is 'mean'.}
}
\value{
Returns a numeric score based on the specified aggregation method.
}
\description{
Function to aggregate geneList based on geneSet, evaluating one geneSet at a time.
The function supports multiple aggregation methods as specified by the user.
}
