% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate_geneSet.R
\name{aggregate_geneSetList}
\alias{aggregate_geneSetList}
\title{Aggregate Gene Set List}
\usage{
aggregate_geneSetList(geneSetList, geneList, ...)
}
\arguments{
\item{geneSetList}{A list of multiple gene sets of interest after filtering.}

\item{geneList}{A list of genes, which can be either a true gene list or a null gene list.}

\item{...}{Additional arguments passed to the \code{aggregate_geneSet} function.}
}
\value{
A list of aggregated scores for each gene set.
}
\description{
This function aggregates multiple gene sets of interest after filtering, using a provided gene list.
}
