% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{anno_data_synGO}
\alias{anno_data_synGO}
\title{Annotation Data for SynGO}
\format{
A data frame with various annotation data.
}
\source{
BrainEnrich::get_annoData(type='SynGO')
}
\usage{
anno_data_synGO
}
\description{
This dataset contains annotation data fetched from the BrainEnrich package for SynGO.
}
\keyword{datasets}
