% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{annoData_synGO}
\alias{annoData_synGO}
\title{SynGO Annotation Data}
\format{
A data frame with various annotation data.
}
\source{
BrainEnrich::get_annoData(type = 'SynGO')
}
\usage{
annoData_synGO
}
\description{
This dataset contains annotation data fetched from the BrainEnrich package for SynGO.
}
\keyword{datasets}
