% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{brain_data_dk_lh_Cushing}
\alias{brain_data_dk_lh_Cushing}
\title{Brain Data Desikan Left Hemisphere Cushing}
\format{
A data frame with rows as regions and columns as Cohen's d values.
}
\source{
read.csv('data-raw/desikan_cushing_data.csv')
}
\usage{
brain_data_dk_lh_Cushing
}
\description{
This dataset contains Cushing's data, filtered and reordered based on PC1 data for regions starting with 'L_' in the Desikan atlas.
}
\keyword{datasets}
