% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{brain_data_dk_lh_PC1}
\alias{brain_data_dk_lh_PC1}
\title{Brain Data Desikan Left Hemisphere PC1}
\format{
A data frame with rows as regions and columns as PC1 data.
}
\source{
read.csv('data-raw/desikan_PC1_data.csv')
}
\usage{
brain_data_dk_lh_PC1
}
\description{
This dataset contains PC1 data filtered for regions starting with 'L_' in the Desikan atlas.
}
\keyword{datasets}
