% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{brain_data_random}
\alias{brain_data_random}
\title{Random Desikan region for testing brainscore function}
\format{
Simulated data frame for 34 regions by 100 subject.
}
\source{
read.csv('data-raw/desikan_centroid.csv')
}
\usage{
brain_data_random
}
\description{
Random Desikan region for testing brainscore function
}
\keyword{datasets}
