% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brainscore.R
\name{brainscore}
\alias{brainscore}
\title{Calculate Brain Scores for Gene Sets}
\usage{
brainscore(
  brain_data,
  gene_data,
  annoData,
  cor_method = c("pearson", "spearman", "pls1c", "pls1w", "custom"),
  aggre_method = c("mean", "median", "meanabs", "meansqr", "maxmean", "ks_orig",
    "ks_weighted", "ks_pos_neg_sum", "local_fdr", "sign_test", "rank_sum", "custom"),
  prefix = NULL,
  minGSSize = 10,
  maxGSSize = 200
)
}
\arguments{
\item{brain_data}{A data frame of brain data. Region by 1 column.}

\item{gene_data}{A data frame of gene expression data.}

\item{annoData}{An environment containing annotation data.}

\item{cor_method}{A character string specifying the correlation method.
Default is 'pearson'. Other options include 'spearman', 'pls1c',
'pls1w', 'custom'.}

\item{aggre_method}{A character string specifying the aggregation method.
Default is 'mean'. Other options include 'median', 'meanabs',
'meansqr', 'maxmean', 'ks_orig', 'ks_weighted', 'ks_pos_neg_sum',
'local_fdr', 'sign_test', 'rank_sum', 'custom'.}

\item{prefix}{A character string to be prefixed to the column names of the result. Default is NULL.}

\item{minGSSize}{An integer specifying the minimum gene set size. Default is 10.}

\item{maxGSSize}{An integer specifying the maximum gene set size. Default is 200.}
}
\value{
A data frame containing the gene set scores with regions as rows and gene sets as columns.
}
\description{
This function calculates scores for gene sets based on brain data.
}
