% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{caculate_pvals}
\alias{caculate_pvals}
\title{Calculate correlations or associations between gene and brain data}
\usage{
caculate_pvals(
  statList.true,
  statList.null,
  method = c("standard", "split_pos_neg")
)
}
\arguments{
\item{statList.true}{A data frame or matrix of gene expression data.}

\item{statList.null}{A data frame or matrix of brain data.}

\item{method}{The method to be used for correlation/association. Can be 'pearson', 'spearman',
'pls1c', 'pls1w', or a custom function provided by the user.}
}
\value{
A matrix with correlation or association coefficients between gene data and brain data.
}
\description{
Calculate correlations or associations between gene and brain data
}
\examples{
# Custom correlation function example
custom_cor <- function(gene_data, brain_data) {
  apply(gene_data, 2, function(g) apply(brain_data, 2, function(b) cor(g, b, method = "pearson")))
}
result <- corr_brain_gene(gene_data, brain_data, method = custom_cor)
}
