% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{calculate_qvalue}
\alias{calculate_qvalue}
\title{Calculate Q-Values}
\usage{
calculate_qvalue(pvals)
}
\arguments{
\item{pvals}{A numeric vector of p-values.}
}
\value{
A numeric vector of q-values corresponding to the input p-values, or NA if q-value calculation fails.
}
\description{
This function calculates q-values from a vector of p-values using the \code{qvalue} package.
Q-values are used to control the false discovery rate (FDR) in multiple hypothesis testing.
}
\examples{
pvals <- runif(100)
qvals <- calculate_qvalue(pvals)
print(qvals)
}
