% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{coord_dk_lh}
\alias{coord_dk_lh}
\title{Desikan Centroid Coordinates for Left Hemisphere}
\format{
A data frame with rows as regions and columns as coordinates (x, y, z).
}
\source{
read.csv('data-raw/desikan_centroid.csv')
}
\usage{
coord_dk_lh
}
\description{
This dataset contains the centroid coordinates for the Desikan atlas regions in the left hemisphere.
}
\keyword{datasets}
