% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Cushing_data}
\alias{Cushing_data}
\title{Cushing's Disease Related Data (Left Hemisphere)}
\format{
A \code{\link[data.frame]{data.frame}} with regions as row names.
}
\source{
Adapted from clinical study data on Cushing's disease, aligned with Desikan atlas regions (left hemisphere).
}
\usage{
Cushing_data
}
\description{
Processed data from a study on Cushing's disease, focusing on Cohen's d values for brain regions,
specifically ordered and filtered for the left hemisphere using the Desikan atlas.
}
\examples{
data(Cushing_data)
plot(Cushing_data$cohend, type = 'h')
}
\keyword{datasets}
