% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_geneSets.R
\name{entrezid2symbol}
\alias{entrezid2symbol}
\title{Convert Entrez IDs to Gene Symbols}
\usage{
entrezid2symbol(entrezid)
}
\arguments{
\item{entrezid}{A vector of Entrez IDs to be converted to gene symbols.}
}
\value{
A vector of gene symbols corresponding to the input Entrez IDs.
}
\description{
This function converts a vector of Entrez IDs to gene symbols using the org.Hs.eg.db annotation package.
}
