% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_geneSets.R
\name{filter_geneSetList}
\alias{filter_geneSetList}
\title{Filter Gene Set List}
\usage{
filter_geneSetList(bg_genes, geneSetList, minGSSize, maxGSSize)
}
\arguments{
\item{bg_genes}{A vector of background gene symbols to be used for filtering.}

\item{geneSetList}{A list of gene sets to be filtered.}

\item{minGSSize}{Minimum gene set size for filtering.}

\item{maxGSSize}{Maximum gene set size for filtering.}
}
\value{
A filtered list of gene sets that meet the size constraints and background genes criteria.
}
\description{
This function filters a list of gene sets based on the background genes and specified size constraints.
}
