% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_core_genes.R
\name{find_core_genes}
\alias{find_core_genes}
\title{Find Core Genes Influencing Aggregated Score}
\usage{
find_core_genes(
  geneList,
  geneSetList,
  method,
  n_cores = 1,
  threshold_type = c("sd", "percentile"),
  threshold = 1
)
}
\arguments{
\item{geneList}{A matrix of genes by models, each column representing a model.}

\item{geneSetList}{A list of gene sets, each containing names of genes.}

\item{method}{The aggregation method used to compute the scores.}

\item{n_cores}{The number of cores to use for parallel processing; defaults to 1.
Uses all available cores minus one if set to 0.}

\item{threshold_type}{The method to determine significance ('sd' for standard deviation, 'percentile' for percentile threshold).}

\item{threshold}{Numeric value specifying the threshold level; meaning depends on \code{threshold_type}.}
}
\value{
A list of core genes for each gene set.
}
\description{
This function performs a Leave-One-Out (LOO) analysis on gene sets to determine core genes
that significantly influence the aggregated score. It utilizes parallel processing
to enhance computation efficiency.
}
\examples{
geneList <- matrix(rnorm(100), ncol = 10, dimnames = list(gene = paste("Gene", 1:10), model = 1:10))
geneSetList <- list(gene_set1 = paste("Gene", sample(1:10, 5), sep = ""))
core_genes <- find_core_genes(geneList, geneSetList, method = 'mean', n_cores = 2, threshold_type = 'sd', threshold = 1)
}
