% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gene_data_dk_lh}
\alias{gene_data_dk_lh}
\title{Gene Expression Data Desikan Left Hemisphere}
\format{
A data frame with various gene expression data.
}
\source{
BrainEnrich::get_geneExp(atlas = 'desikan', rdonor = 'r0.6', hem = 'L')
}
\usage{
gene_data_dk_lh
}
\description{
This dataset contains gene expression data fetched from the BrainEnrich package for the Desikan atlas, left hemisphere.
}
\keyword{datasets}
