% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spin_brain_functions.R
\name{generate_null_brain_data}
\alias{generate_null_brain_data}
\title{Generate Null Brain Data}
\usage{
generate_null_brain_data(brain_data, perm_id)
}
\arguments{
\item{brain_data}{A matrix representing brain data, where each row corresponds to a region.}

\item{perm_id}{A matrix of permutations, where each column represents a permutation and each row corresponds to an index in \code{brain_data}.}
}
\value{
A matrix of null brain data with the same dimensions as \code{brain_data} but with permuted rows according to \code{perm_id}.
}
\description{
This function generates null brain datasets based on permutations provided in perm_id.
It rearranges the brain data according to the permutations and outputs the shuffled datasets.
}
