% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_brainExample .R
\name{get_brainExample}
\alias{get_brainExample}
\title{Load example Data}
\usage{
get_brainExample(type = c("PC1", "Cushing"))
}
\arguments{
\item{filename}{Character. The name of the file to load.}
}
\value{
A data frame containing the example data.
}
\description{
This function loads example data included with the package.
}
\examples{
example_data <- load_example_data('example.csv')
}
