% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_geneSets.R
\name{get_celltype_data}
\alias{get_celltype_data}
\title{Download and process cell type gene set data}
\usage{
get_celltype_data(type = c("Seidlitz2020", "Lake2018", "Martins2021"))
}
\arguments{
\item{type}{The type of cell type gene set data to download and process. Must be one of 'Seidlitz2020', 'Lake2018', 'Martins2021'.}
}
\value{
A list of gene sets where each element is a vector of genes associated with a specific term.
}
\description{
This function downloads cell type gene set data from specified resources, processes the data,
and returns a list of gene sets based on the specified type.
}
\details{
The function performs the following steps based on the specified type:
\itemize{
\item Downloads the cell type gene set data from the provided URL.
\item Reads the data from the file and processes it to create a gene set list.
\item Supports the following types:
\itemize{
\item \strong{Seidlitz2020}: \url{https://github.com/jms290/PolySyn_MSNs/blob/master/Data/AHBA/celltypes_PSP.csv}
\itemize{
\item Seidlitz, J., Nadig, A., Liu, S., Bethlehem, R. A., Vértes, P. E., Morgan, S. E., ... & Raznahan, A. (2020). Transcriptomic and cellular decoding of regional brain vulnerability to neurogenetic disorders. Nature communications, 11(1), 3358.
}
\item \strong{Lake2018}: \url{https://github.com/molecular-neuroimaging/Imaging_Transcriptomics/blob/main/imaging_transcriptomics/data/geneset_LAKE.gmt}
\itemize{
\item Lake, B. B., Chen, S., Sos, B. C., Fan, J., Kaeser, G. E., Yung, Y. C., ... & Zhang, K. (2018). Integrative single-cell analysis of transcriptional and epigenetic states in the human adult brain. Nature biotechnology, 36(1), 70-80.
}
\item \strong{Martins2021}: \url{https://github.com/molecular-neuroimaging/Imaging_Transcriptomics/blob/main/imaging_transcriptomics/data/geneset_Pooled.gmt}
\itemize{
\item Imaging transcriptomics: Convergent cellular, transcriptomic, and molecular neuroimaging signatures in the healthy adult human brain. Daniel Martins, Alessio Giacomel, Steven CR Williams, Federico Turkheimer, Ottavia Dipasquale, Mattia Veronese, PET templates working group. Cell Reports.
}
}
}
}
\examples{
\dontrun{
# Example usage of get_celltype_data function
geneSetList <- get_celltype_data('Seidlitz2020')
print(geneSetList)
}

}
