% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_geneExp.R
\name{get_geneExp}
\alias{get_geneExp}
\title{Get Gene Expression Data}
\usage{
get_geneExp(
  atlas = c("desikan", "schaefer100", "schaefer200", "schaefer300"),
  rdonor = c("r0.2", "r0.4", "r0.6"),
  hem = c("L", "R", "B")
)
}
\arguments{
\item{atlas}{Character. The brain atlas to use. One of 'desikan', 'schaefer100', or 'schaefer200'.}

\item{rdonor}{Character. The donor region to use. One of 'r0.2', 'r0.4', or 'r0.6'.}

\item{hem}{Character. The hemisphere to use. One of 'L', 'R', or 'B'.}
}
\value{
A matrix containing the gene expression data.
}
\description{
This function retrieves gene expression data based on specified parameters.
The data is obtained from the ENIGMA-TOOLBOX. Please cite the ENIGMA-TOOLBOX:
Larivière, S., Paquola, C., Park, B. Y., Royer, J., Wang, Y., Benkarim, O., ... & Bernhardt, B. C. (2021).
The ENIGMA Toolbox: multiscale neural contextualization of multisite neuroimaging datasets. Nature Methods, 18(7), 698-700.
}
\examples{
gene_exp_matrix <- get_geneExp(atlas = 'desikan', rdonor = 'r0.4', hem = 'L')
}
