% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_geneSets.R
\name{get_geneSetDescription}
\alias{get_geneSetDescription}
\title{Get Gene Set Descriptions}
\usage{
get_geneSetDescription(annoData)
}
\arguments{
\item{annoData}{An environment containing annotation data.}
}
\value{
A character vector of gene set descriptions.
}
\description{
This function retrieves descriptions for gene sets from annotation data.
}
