% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_geneSets.R
\name{get_geneSetList}
\alias{get_geneSetList}
\title{Get Gene Set List}
\usage{
get_geneSetList(annoData, convert_to_symbol = FALSE)
}
\arguments{
\item{annoData}{Annotation data to retrieve gene sets from.}

\item{convert_to_symbol}{Logical; if TRUE, converts gene identifiers to gene symbols.}
}
\value{
A list of gene sets.
}
\description{
This function retrieves a gene set list from annotation data. It optionally converts
gene identifiers to gene symbols.
}
