% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_geneSets.R
\name{get_SynGO_data}
\alias{get_SynGO_data}
\title{Download and process SynGO data}
\usage{
get_SynGO_data(
  url = "https://www.syngoportal.org/data/SynGO_bulk_download_release_20231201.zip"
)
}
\arguments{
\item{url}{The URL to download the SynGO data zip file from. Default is the latest release URL.}
}
\value{
A data frame containing the processed SynGO data.
}
\description{
This function downloads the SynGO data from the specified URL, extracts the \code{syngo_ontologies.xlsx} file,
processes the data to create TERM2GENE and TERM2NAME data frames, builds the annotation using \code{build_Anno},
and removes the downloaded and temporary files afterward.
}
\details{
The function performs the following steps:
\itemize{
\item Downloads the SynGO data zip file from the provided URL.
\item Extracts the \code{syngo_ontologies.xlsx} file from the zip archive.
\item Reads the Excel file and processes the data to create TERM2GENE and TERM2NAME data frames.
\item Uses the \code{build_Anno} function to build the final annotation data.
\item Cleans up by removing the temporary zip file and extracted Excel file.
}
}
