% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_core_genes.R
\name{identify_core_genes_percentile}
\alias{identify_core_genes_percentile}
\title{Identify core genes based on a percentile threshold}
\usage{
identify_core_genes_percentile(changes, percentile = 90)
}
\arguments{
\item{changes}{Named vector of changes from LOO analysis.}

\item{percentile}{Numeric value indicating the percentile to use as the threshold.}
}
\value{
Vector of core genes.
}
\description{
Identify core genes based on a percentile threshold
}
