% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_core_genes.R
\name{identify_core_genes_sd}
\alias{identify_core_genes_sd}
\title{Identify core genes based on a standard deviation threshold}
\usage{
identify_core_genes_sd(changes, k = 1)
}
\arguments{
\item{changes}{Named vector of changes from LOO analysis.}

\item{k}{Numeric value indicating how many standard deviations to use as the threshold.}
}
\value{
Vector of core genes.
}
\description{
Identify core genes based on a standard deviation threshold
}
