% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{PC1_data}
\alias{PC1_data}
\title{Principal Component 1 Data Filtered by Region (Left Hemisphere)}
\format{
A \code{\link[data.frame]{data.frame}} with regions as row names.
}
\source{
Processed based on PC1 analysis from MRI data of the Desikan atlas (left hemisphere).
}
\usage{
PC1_data
}
\description{
The first principal component of brain data from the Desikan atlas,
filtered to include only regions starting with 'L_', indicative of the left hemisphere.
}
\examples{
data(PC1_data)
head(PC1_data)
}
\keyword{datasets}
