% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{perm_id}
\alias{perm_id}
\title{Permutation ID Data for Desikan Spin Brain (Left Hemisphere)}
\format{
An R object loaded via \code{\link[base]{readRDS}}
}
\source{
Data generated internally using Desikan atlas configurations.
}
\usage{
perm_id
}
\description{
Permutation identifiers used for analyses involving spatial autocorrelation in neuroimaging data,
specifically tailored for the left hemisphere based on the Desikan atlas.
}
\examples{
data(perm_id)
length(perm_id)
}
\keyword{datasets}
