% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{perm_id_dk_lh}
\alias{perm_id_dk_lh}
\title{Permutation IDs for Desikan Left Hemisphere}
\format{
A matrix with permutation IDs for regions.
}
\source{
rotate_parcellation(coord.l = coord_dk_lh, nrot = 1000, seed = 2024)
}
\usage{
perm_id_dk_lh
}
\description{
This dataset contains permutation IDs generated for the Desikan spin brain for the left hemisphere.
}
\keyword{datasets}
