% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resample_gene_functions.R
\name{resample_gene}
\alias{resample_gene}
\title{Resample Gene List}
\usage{
resample_gene(geneList.true, n_perm = 5000)
}
\arguments{
\item{geneList.true}{A matrix of gene expression data.}

\item{n_perm}{Number of permutations to generate.}
}
\value{
A matrix of permuted gene lists.
}
\description{
Generates a set of permuted gene lists from the original gene list,
ensuring uniqueness in the permuted sets.
}
