% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/competitive_null_models.R
\name{resampling_geneList}
\alias{resampling_geneList}
\title{Resample Gene List}
\usage{
resampling_geneList(geneList.true, perm.n = 5000)
}
\arguments{
\item{geneList.true}{A matrix of gene expression data.}

\item{perm.n}{Number of permutations to generate.}
}
\value{
A matrix of permuted gene lists.
}
\description{
Generates a set of permuted gene lists from the original gene list,
ensuring uniqueness in the permuted sets.
}
\examples{
gene_data <- matrix(rnorm(100), nrow=10)
resampled_gene_list <- resampling_geneList(gene_data, perm.n=100)
}
