% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/competitive_null_models.R
\name{resampling_geneSetList_with_constraints}
\alias{resampling_geneSetList_with_constraints}
\title{Resample Gene Sets with Specified Constraints}
\usage{
resampling_geneSetList_with_constraints(
  geneSetList,
  constrain = c("match_coexp", "gene_subset"),
  coexp_matrix = NULL,
  gene_subset = NULL
)
}
\arguments{
\item{geneSetList}{A list of gene sets to be resampled.}

\item{constrain}{A character vector specifying the type of constraint.
Available options are 'match_coexp' for matching co-expression and
'gene_subset' for resampling within a specified subset of genes.}

\item{coexp_matrix}{An optional co-expression matrix, required if 'match_coexp' is selected.}

\item{gene_subset}{An optional vector of gene identifiers, required if 'gene_subset' is selected.}
}
\value{
A list of resampled gene sets based on the specified constraints.
}
\description{
This function resamples gene sets based on specific constraints like matching
co-expression patterns or restricting to a gene subset.
The methodology implemented is informed by Wei et al. (2022) on statistical testing in transcriptomic-neuroimaging studies.
It is important to note that restricting null models to a subset of genes can be problematic. The empirical statistics sampled from
the full gene pool differ from those derived from a restricted pool. Therefore, usage of this approach should be
with caution.
}
\examples{
# Assuming geneSetList, coexp_matrix, and gene_subset are predefined:
resampled_sets_coexp = resampling_geneSetList_with_constraints(geneSetList,
                                     constrain='match_coexp', coexp_matrix=coexp_matrix)
resampled_sets_subset = resampling_geneSetList_with_constraints(geneSetList,
                                     constrain='gene_subset', gene_subset=gene_subset)
}
\references{
Wei, Y., de Lange, S. C., Pijnenburg, R., Scholtens, L. H., Ardesch, D. J., Watanabe, K., Posthuma, D., & van den Heuvel, M. P. (2022).
Statistical testing in transcriptomic-neuroimaging studies: A how-to and evaluation of methods assessing spatial and gene specificity.
Human Brain Mapping, 43(3), 885–901. \url{https://doi.org/10.1002/hbm.25711}
}
