% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spin_brain_functions.R
\name{rotate_parcellation}
\alias{rotate_parcellation}
\title{Rotate Parcellation}
\usage{
rotate_parcellation(
  coord.l = NULL,
  coord.r = NULL,
  nrot = 5000,
  method = c("hungarian", "vasa"),
  seed = NULL
)
}
\arguments{
\item{coord.l}{Coordinates of left hemisphere regions on the sphere (array of size n(LH regions) x 3). Can be NULL if only right hemisphere is used.}

\item{coord.r}{Coordinates of right hemisphere regions on the sphere (array of size n(RH regions) x 3). Can be NULL if only left hemisphere is used.}

\item{nrot}{Number of rotations (default = 5000).}

\item{method}{Method to match rotated and unrotated regions; options are 'vasa' (faster, can be suboptimal) or 'hungarian' (default, slower, optimal).}
}
\value{
Array of permutations, from set of regions to itself (array of size n(total regions) x nrot).
}
\description{
Generate a permutation map from a set of cortical regions of interest to itself,
while (approximately) preserving contiguity and hemispheric symmetry.
The function is based on a rotation of the FreeSurfer projection of coordinates
of a set of regions of interest on the sphere.
#' This function is modified from the original version available at:
https://github.com/frantisekvasa/rotate_parcellation
}
\details{
Modifications include:
\itemize{
\item Added support for scenarios where only one hemisphere's coordinates are provided.
\item Improved handling of coordinate dimensions and conditional concatenation of reference and rotation indices.
\item Included importFrom directives for required functions from \code{matrixStats} and \code{clue}.
\item Ensured the function generates \code{nrot + 100} permutations, removes duplicates, and returns exactly \code{nrot} unique permutations.
}
}
\examples{
# Example usage with both hemispheres
coord.l <- matrix(runif(30), nrow = 10, ncol = 3)
coord.r <- matrix(runif(30), nrow = 10, ncol = 3)
permutations <- rotate_parcellation(coord.l, coord.r)

# Example usage with one hemisphere (left)
coord.l <- matrix(runif(30), nrow = 10, ncol = 3)
permutations <- rotate_parcellation(coord.l, NULL)

# Example usage with one hemisphere (right)
coord.r <- matrix(runif(30), nrow = 10, ncol = 3)
permutations <- rotate_parcellation(NULL, coord.r)
}
