% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resample_gene_functions.R
\name{sample_gs_matching_coexp}
\alias{sample_gs_matching_coexp}
\title{Sample Gene Sets Matching Co-Expression}
\usage{
sample_gs_matching_coexp(
  gs,
  coexp_matrix,
  tol = 0.01,
  max_iter = 1e+06,
  n_target = 5000
)
}
\arguments{
\item{gs}{The target gene set for which similar co-expression profiles are sought.}

\item{coexp_matrix}{A co-expression matrix, typically calculated as the correlation matrix of gene expression data.}

\item{tol}{Tolerance for the difference between the co-expression of the target and the sampled gene sets.}

\item{max_iter}{Maximum number of iterations to attempt finding matches.}

\item{n_target}{Number of gene sets to sample.}
}
\value{
A list of gene sets that closely match the target gene set's co-expression profile.
}
\description{
This function samples gene sets that closely match the co-expression profile of a target gene set.
The methodology implemented is informed by Wei et al. (2022) on statistical testing in transcriptomic-neuroimaging studies.
}
\references{
Wei, Y., de Lange, S. C., Pijnenburg, R., Scholtens, L. H., Ardesch, D. J., Watanabe, K., Posthuma, D., & van den Heuvel, M. P. (2022).
Statistical testing in transcriptomic-neuroimaging studies: A how-to and evaluation of methods assessing spatial and gene specificity.
Human Brain Mapping, 43(3), 885–901. \url{https://doi.org/10.1002/hbm.25711}
Assuming 'coexp_matrix' is predefined:
target_gene_set <- c("Gene1", "Gene2", "Gene3")
similar_gene_sets <- sample_gs_matching_coexp(target_gene_set, coexp_matrix)
}
