% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/competitive_null_models.R
\name{sample_gs_within_subset}
\alias{sample_gs_within_subset}
\title{Sample Gene Sets Within a Specified Subset}
\usage{
sample_gs_within_subset(gs, gene_subset, perm.n = 5000)
}
\arguments{
\item{gs}{The original gene set for reference size.}

\item{gene_subset}{A vector of gene identifiers from which to sample.}

\item{perm.n}{The number of gene sets to generate.}
}
\value{
A list of gene sets sampled from the specified subset.
}
\description{
This function samples gene sets by randomly selecting genes within a specified subset.
The methodology implemented is informed by Wei et al. (2022) on statistical testing in transcriptomic-neuroimaging studies.
It is important to note that restricting null models to a subset of genes can be problematic. The empirical statistics sampled from
the full gene pool differ from those derived from a restricted pool. Therefore, usage of this approach should be
with caution.
}
\references{
Wei, Y., de Lange, S. C., Pijnenburg, R., Scholtens, L. H., Ardesch, D. J., Watanabe, K., Posthuma, D., & van den Heuvel, M. P. (2022).
Statistical testing in transcriptomic-neuroimaging studies: A how-to and evaluation of methods assessing spatial and gene specificity.
Human Brain Mapping, 43(3), 885–901. \url{https://doi.org/10.1002/hbm.25711}
}
