% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resample_gene_functions.R
\name{swap_geneList}
\alias{swap_geneList}
\title{Swap Gene List}
\usage{
swap_geneList(geneList.true, orig_gs, sampled_gs)
}
\arguments{
\item{geneList.true}{A matrix representing the true gene list, with dimensions m x 1.}

\item{orig_gs}{A vector of original gene set identifiers.}

\item{sampled_gs}{A list of sampled gene sets, each being a vector of gene identifiers.}
}
\value{
A matrix where each column represents a null gene list generated by swapping \code{orig_gs} with each set in \code{sampled_gs}.
}
\description{
This function swaps the original values of a gene set (\code{orig_gs}) with the sampled gene sets (\code{sampled_gs}) within the given \code{geneList.true}.
The resulting \code{geneList.null} is in the same format as generated by other approaches.
}
\examples{
geneList.true <- matrix(1:10, ncol = 1)
rownames(geneList.true) <- letters[1:10]
orig_gs <- c("a", "b", "c")
sampled_gs <- list(c("d", "e", "f"), c("g", "h", "i"))
swap_geneList(geneList.true, orig_gs, sampled_gs)
}
