% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{syngo_annoData}
\alias{syngo_annoData}
\title{SynGO Annotation Data}
\format{
A \code{\link[data.frame]{data.frame}}
}
\source{
SynGO project, sourced from \url{https://www.syngoportal.org/}
}
\usage{
syngo_annoData
}
\description{
This dataset includes synaptic functionalities and gene ontology annotations from the SynGO project,
useful for studies focused on synaptic functions and disorders.
}
\examples{
data(syngo_annoData)
summary(syngo_annoData)
}
\keyword{datasets}
