% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_statistics.R
\name{aggregate_geneSet}
\alias{aggregate_geneSet}
\title{Aggregate Gene List Based on a Gene Set}
\usage{
aggregate_geneSet(
  geneList,
  geneSet,
  method = c("mean", "median", "meanabs", "meansqr", "maxmean", "sig_n", "sign_test",
    "rank_sum", "ks_orig", "ks_weighted", "ks_sum", "locfdr")
)
}
\arguments{
\item{geneList}{A matrix representing named correlation coefficients, with each column representing a model.}

\item{geneSet}{A vector of gene names representing the gene set of interest.}

\item{method}{A character string specifying the aggregation method to be used, or a custom function provided by the user.}
}
\value{
Returns a numeric vector of aggregated gene scores.
}
\description{
This function aggregates a gene list based on a specified gene set using various methods.
}
\examples{
geneList <- matrix(rnorm(100), nrow = 10)
rownames(geneList) <- paste("Gene", 1:10, sep = "")
geneSet <- rownames(geneList)[1:5]
aggregate_geneSet(geneList, geneSet, method = "mean")
}
