% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate_functions.R
\name{aggregate_geneSetList}
\alias{aggregate_geneSetList}
\title{Aggregate Gene Set List in Parallel}
\usage{
aggregate_geneSetList(geneList, geneSetList, method, n_cores = 1)
}
\arguments{
\item{geneList}{A list of genes.}

\item{geneSetList}{A list of gene sets to be aggregated.}

\item{n_cores}{Number of cores to use for parallel processing. Default is 1.
If set to 0, it uses all available cores minus one.}

\item{...}{Additional arguments to be passed to the \code{aggregate_geneSet} function.}
}
\value{
A list of aggregated gene set scores.
}
\description{
This function aggregates gene sets in parallel using the \code{parLapply} function
from the \code{parallel} package, ensuring cross-platform compatibility.
}
