% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate_functions.R
\name{aggregate_geneSetList_matching_coexp}
\alias{aggregate_geneSetList_matching_coexp}
\title{Aggregate Gene Set List with Matching Coexpression in Parallel}
\usage{
aggregate_geneSetList_matching_coexp(
  geneList.true,
  geneSetList,
  sampled_geneSetList,
  method,
  n_cores = 1
)
}
\arguments{
\item{geneList.true}{A m x 1 matrix of true gene sets. Ensure to include \code{drop=FALSE} when subsetting.}

\item{geneSetList}{A list of gene sets.}

\item{sampled_geneSetList}{A list of sampled gene sets.}

\item{method}{The method to be used for aggregation.}

\item{n_cores}{Number of cores to use for parallel processing. Default is 1.
If set to 0, it uses all available cores minus one.}
}
\value{
A list of aggregated gene set scores.
}
\description{
This function aggregates gene set scores in parallel using \code{pblapply} from the \code{pbapply} package.
}
