% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/self_contained_null_models.R
\name{generate_null_brain_data}
\alias{generate_null_brain_data}
\title{Generate Null Brain Data}
\usage{
generate_null_brain_data(brain_data, perm_id)
}
\arguments{
\item{brain_data}{A matrix representing brain data, where each row corresponds to a region.}

\item{perm_id}{A matrix of permutations, where each column represents a permutation and each row corresponds to an index in \code{brain_data}.}
}
\value{
A matrix of null brain data with the same dimensions as \code{brain_data} but with permuted rows according to \code{perm_id}.
}
\description{
This function generates null brain datasets based on permutations provided in perm.id.
It rearranges the brain data according to the permutations and outputs the shuffled datasets.
}
\examples{
brain_data <- matrix(rnorm(100), nrow=10)
perm_id <- matrix(sample(1:10), nrow=10, ncol=5)
null_data <- generate_null_brain_data(brain_data, perm_id)
}
